# logistics/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "DriverRead": {
        "additionalProperties": false,
        "description": "Persisted driver representation.",
        "properties": {
          "active": {
            "default": true,
            "title": "Active",
            "type": "boolean"
          },
          "external_worker_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Worker Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/NonBlank"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "preferred_route_group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Route Group"
          },
          "scenario_id": {
            "format": "uuid",
            "title": "Scenario Id",
            "type": "string"
          }
        },
        "required": [
          "name",
          "id",
          "scenario_id"
        ],
        "title": "DriverRead",
        "type": "object"
      },
...
      "DriverUpdate": {
        "additionalProperties": false,
        "description": "Partial driver update.",
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "external_worker_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Worker Id"
          },
          "name": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NonBlank"
              },
              {
                "type": "null"
              }
            ]
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "preferred_route_group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Route Group"
          }
        },
        "title": "DriverUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NonBlank": {
        "minLength": 1,
        "type": "string"
      },
...
      "RwmsAppliedAssignment": {
        "additionalProperties": false,
        "description": "Successfully applied or replayed RWMS assignment outcome.",
        "properties": {
          "documentId": {
            "format": "uuid",
            "title": "Documentid",
            "type": "string"
          },
          "orderId": {
            "format": "uuid",
            "title": "Orderid",
            "type": "string"
          },
          "replayed": {
            "title": "Replayed",
            "type": "boolean"
          }
        },
        "required": [
          "orderId",
          "documentId",
          "replayed"
        ],
        "title": "RwmsAppliedAssignment",
        "type": "object"
      },
      "RwmsApplyResult": {
        "additionalProperties": false,
        "description": "Complete upstream application result, including every rejection.",
        "properties": {
          "applied": {
            "items": {
              "$ref": "#/components/schemas/RwmsAppliedAssignment"
            },
            "title": "Applied",
            "type": "array"
          },
          "rejected": {
            "items": {
              "$ref": "#/components/schemas/RwmsRejectedAssignment"
            },
            "title": "Rejected",
            "type": "array"
          }
        },
        "title": "RwmsApplyResult",
        "type": "object"
      },
      "RwmsPlanApplyRequest": {
        "additionalProperties": false,
        "description": "Optimistic plan version plus explicitly selected future shared delivery parts.",
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "publish_unassigned_task_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 500,
            "title": "Publish Unassigned Task Ids",
            "type": "array"
          }
        },
        "required": [
          "expected_version"
        ],
        "title": "RwmsPlanApplyRequest",
        "type": "object"
      },
      "RwmsPlanStatusResult": {
        "additionalProperties": false,
        "description": "Exact plan-version RWMS status snapshot exposed to the operator UI.",
        "properties": {
          "plan_id": {
            "format": "uuid",
            "title": "Plan Id",
            "type": "string"
          },
          "plan_version": {
            "minimum": 1.0,
            "title": "Plan Version",
            "type": "integer"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/RwmsPlanTaskStatus"
            },
            "title": "Tasks",
            "type": "array"
          }
        },
        "required": [
          "plan_id",
          "plan_version"
        ],
        "title": "RwmsPlanStatusResult",
        "type": "object"
      },
...
      "RwmsRejectedAssignment": {
        "additionalProperties": false,
        "description": "RWMS assignment rejection retained for operator action.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "orderId": {
            "format": "uuid",
            "title": "Orderid",
            "type": "string"
          }
        },
        "required": [
          "orderId",
          "code",
          "message"
        ],
        "title": "RwmsRejectedAssignment",
        "type": "object"
      },
...
      "RwmsSyncRequest": {
        "additionalProperties": false,
        "description": "Date-bounded synchronization command for one RWMS warehouse.",
        "properties": {
          "date_from": {
            "format": "date",
            "title": "Date From",
            "type": "string"
          },
          "date_to": {
            "format": "date",
            "title": "Date To",
            "type": "string"
          },
          "warehouse_id": {
            "format": "uuid",
            "title": "Warehouse Id",
            "type": "string"
          }
        },
        "required": [
          "warehouse_id",
          "date_from",
          "date_to"
        ],
        "title": "RwmsSyncRequest",
        "type": "object"
      },
      "RwmsSyncResult": {
        "additionalProperties": false,
        "description": "Counted synchronization result with explicit per-order failures.",
        "properties": {
          "failures": {
            "items": {
              "$ref": "#/components/schemas/RwmsSyncFailure"
            },
            "title": "Failures",
            "type": "array"
          },
          "imported": {
            "title": "Imported",
            "type": "integer"
          },
          "skipped": {
            "title": "Skipped",
            "type": "integer"
          },
          "updated": {
            "title": "Updated",
            "type": "integer"
          }
        },
        "required": [
          "imported",
          "updated",
          "skipped"
        ],
        "title": "RwmsSyncResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WarehouseRead": {
        "additionalProperties": false,
        "description": "Persisted warehouse representation.",
        "properties": {
          "external_warehouse_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Warehouse Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "latitude": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "type": "number"
          },
          "loading_minutes": {
            "default": 30,
            "minimum": 0.0,
            "title": "Loading Minutes",
            "type": "integer"
          },
          "longitude": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "type": "number"
          },
          "name": {
            "$ref": "#/components/schemas/NonBlank"
          },
          "scenario_id": {
            "format": "uuid",
            "title": "Scenario Id",
            "type": "string"
          },
          "turnaround_minutes": {
            "default": 15,
            "minimum": 0.0,
            "title": "Turnaround Minutes",
            "type": "integer"
          },
          "unloading_minutes": {
            "default": 30,
            "minimum": 0.0,
            "title": "Unloading Minutes",
            "type": "integer"
          },
          "working_day_end": {
            "default": "20:00:00",
            "format": "time",
            "title": "Working Day End",
            "type": "string"
          },
          "working_day_start": {
            "default": "08:00:00",
            "format": "time",
            "title": "Working Day Start",
            "type": "string"
          }
        },
        "required": [
          "name",
          "latitude",
          "longitude",
          "id",
          "scenario_id"
        ],
        "title": "WarehouseRead",
        "type": "object"
      },
      "WarehouseUpdate": {
        "additionalProperties": false,
        "description": "Partial update for depot location and service timings.",
        "properties": {
          "external_warehouse_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Warehouse Id"
          },
          "latitude": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "loading_minutes": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Loading Minutes"
          },
          "longitude": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "name": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NonBlank"
              },
              {
                "type": "null"
              }
            ]
          },
          "turnaround_minutes": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turnaround Minutes"
          },
          "unloading_minutes": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unloading Minutes"
          },
          "working_day_end": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Working Day End"
          },
          "working_day_start": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Working Day Start"
          }
        },
        "title": "WarehouseUpdate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Internal logistics planning, route editing, and simulation API.",
    "title": "RWMS Logistics Simulator",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/drivers/{driver_id}": {
      "delete": {
        "description": "Delete a driver not retained by plan history.",
        "operationId": "delete_driver_api_drivers__driver_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "driver_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Driver Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Driver",
        "tags": [
          "catalog"
        ]
      },
      "patch": {
        "description": "Update a driver.",
        "operationId": "update_driver_api_drivers__driver_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "driver_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Driver Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Driver",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/api/plans/{plan_id}/rwms/apply": {
      "post": {
        "description": "Submit the exact plan version and return all upstream rejections unchanged.",
        "operationId": "apply_rwms_plan_api_plans__plan_id__rwms_apply_post",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RwmsPlanApplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RwmsApplyResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Rwms Plan",
        "tags": [
          "rwms-integration"
        ]
      }
    },
    "/api/plans/{plan_id}/rwms/status": {
      "get": {
        "description": "Return current RWMS publication and claim states for one exact plan version.",
        "operationId": "rwms_plan_status_api_plans__plan_id__rwms_status_get",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plan Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "expected_version",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Expected Version",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RwmsPlanStatusResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rwms Plan Status",
        "tags": [
          "rwms-integration"
        ]
      }
    },
...
    "/api/scenarios/{scenario_id}/rwms/sync": {
      "post": {
        "description": "Import a bounded RWMS delivery feed without geocoding address-only orders.",
        "operationId": "sync_rwms_requests_api_scenarios__scenario_id__rwms_sync_post",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RwmsSyncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RwmsSyncResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Rwms Requests",
        "tags": [
          "rwms-integration"
        ]
      }
    },
...
    "/api/warehouses/{warehouse_id}": {
      "delete": {
        "description": "Delete an unused depot.",
        "operationId": "delete_warehouse_api_warehouses__warehouse_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "warehouse_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Warehouse Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Warehouse",
        "tags": [
          "catalog"
        ]
      },
      "get": {
        "description": "Read a depot by UUID.",
        "operationId": "get_warehouse_api_warehouses__warehouse_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "warehouse_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Warehouse Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Warehouse",
        "tags": [
          "catalog"
        ]
      },
      "patch": {
        "description": "Update a depot.",
        "operationId": "update_warehouse_api_warehouses__warehouse_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "warehouse_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Warehouse Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Warehouse",
        "tags": [
          "catalog"
        ]
      }
    },
...
  }
}